// src/app/pages/home/home/components/models/FilterVisibalModel.ts
export class FilterVisibalModel{
    visibleAmenities: string[] = [];
    visibleAccommodations: string[] = [];
    visibleLocations: string[] = [];
    amenityItems: string[] = [];
    accommodationItems: string[] = [];
    locationItems: string[] = [];
    showMore = 'show more +';
    activeDropdown: number = -1;
    dropdownCount: number = 7;

}